Move saveAsMarkdown
