+++ src/features/textEditor/components/TipTapEditor.tsx
@@ -12,8 +12,7 @@ import { EditorId, parseEditorId } from '../../../atoms/types/EditorData';
 import { Spinner } from '../../../components/Spinner';
 import { emitEvent } from '../../../events';
 import { useListenEvent } from '../../../hooks/useListenEvent';
-import { getFileNameAndExtension, makeExportsPath, writeFileContent } from '../../../io/filesystem';
-import { notifyWarning } from '../../../notifications/notifications';
+import { saveAsMarkdown } from '../saveAsMarkdown';
 import { MenuBar } from './MenuBar';
 import { EDITOR_EXTENSIONS, transformPasted } from './tipTapEditorConfigs';
 import { MarkdownSerializer } from './tipTapNodes/refStudioDocument/serialization/MarkdownSerializer';
@@ -107,27 +106,4 @@ export function TipTapEditor({ editorContent, editorId, isActive, saveFileInMemo
       <EditorContent className="tiptap-editor" editor={editor} />
     </div>
   );
-}
-
-export async function saveAsMarkdown(markdownSerializer: MarkdownSerializer, filePath: string) {
-  try {
-    const { name: fileName } = getFileNameAndExtension(filePath);
-    if (!fileName) {
-      notifyWarning('Could not save markdown file');
-      return;
-    }
-
-    const markdownFileName = `${fileName}.md`;
-    const markdownFilePath = makeExportsPath(markdownFileName);
-    const serializedContent = markdownSerializer.serialize(fileName);
-    await writeFileContent(markdownFilePath, serializedContent.markdownContent);
-
-    if (serializedContent.bibliography) {
-      const bibliographyFileName = `${fileName}.${serializedContent.bibliography.extension}`;
-      const bibliographyFilePath = makeExportsPath(bibliographyFileName);
-      await writeFileContent(bibliographyFilePath, serializedContent.bibliography.textContent);
-    }
-  } catch (err) {
-    console.error('Error', err);
-  }
 }
